Clean up typed ImageData

+++ src/gl/raw/caps.rs
@@ -2,7 +2,7 @@ use glow::HasContext;
 
 #[derive(Debug, Copy, Clone)]
 pub struct Caps {
-    pub max_texture_size: usize,
+    pub max_texture_size: u32,
 }
 
 impl Caps {
@@ -10,7 +10,7 @@ impl Caps {
         let max_texture_size = unsafe { gl.get_parameter_i32(glow::MAX_TEXTURE_SIZE) };
 
         Caps {
-            max_texture_size: usize::try_from(max_texture_size).unwrap(),
+            max_texture_size: u32::try_from(max_texture_size).unwrap(),
         }
     }
 }
